requirements.txt
torch>=2.1.0
transformers>=4.45.0
peft>=0.13.0
bitsandbytes>=0.44.0
accelerate>=0.34.0
datasets>=3.0.0
pyyaml>=6.0
